server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.1",
    "fastapi>=0.129.2",
    "openai>=2.21.0",
    "py-clob-client>=0.34.6",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
